+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/OpenAIService.kt
@@ -20,7 +20,7 @@ class OpenAIService {
     }
 
     private fun getPrompt(locale: String, diff: String) =
-        "Write an insightful but concise Git commit message in a complete sentence in present tense for the following diff without prefacing it with anything, the response must be in the language ${locale}:\\n${diff}"
+        "Write an insightful but concise Git commit message in a complete sentence in present tense for the following diff without prefacing it with anything, the response must be in the language ${locale}. The following text are the differences between files, where deleted lines are prefixed with a single minus sign and added lines are prefixed with a single plus sign:\\n${diff}"
 
     @OptIn(BetaOpenAI::class)
     suspend fun generateCommitMessage(diff: String, completions: Int): String {